+++ niljs/README.md
@@ -3,36 +3,20 @@
 <br />
 
 <p align="center">
-  The TypeScript client library for interacting with the =nil; cluster.
+  The TypeScript/Javascript client library for interacting with the =nil; cluster.
 </p>
 
 <row style="display: flex; gap: 10px;"><p align="center">
-<a href="https://github.com/NilFoundation/nil.js/actions/workflows/build.yaml">
-<picture>
-<img src="https://img.shields.io/github/actions/workflow/status/NilFoundation/nil.js/.github%2Fworkflows%2Fbuild.yaml"/>
-</picture>
-</a>
 <a href="https://www.npmjs.com/package/@nilfoundation/niljs">
 <picture>
 <img src="https://img.shields.io/npm/dy/%40nilfoundation%2Fniljs"/>
 </picture>
 </a>
-<a href="https://github.com/NilFoundation/nil.js">
-<picture>
-<img src="https://img.shields.io/github/stars/NilFoundation/nil.js"/>
-</picture>
-</a>
-<a href="https://github.com/NilFoundation/nil.js/actions/workflows/build.yaml">
+<a href="https://github.com/NilFoundation/nil/actions/workflows/build.yaml">
 <picture>
 <img src="https://img.shields.io/npm/v/%40nilfoundation%2Fniljs"/>
 </picture>
 </a>
-<a href="https://github.com/NilFoundation/nil.js">
-<picture>
-<img src="https://img.shields.io/github/forks/NilFoundation/nil.js"/>
-</picture>
-</a>
-
 </p>
 </row>
 
@@ -83,6 +67,9 @@ const smartAccount = await generateSmartAccount({
   faucetEndpoint: FAUCET_ENDPOINT,
 });
 ```
+
+
+Full docs available here: [https://docs.nil.foundation/nil/niljs/getting-started/](https://docs.nil.foundation/nil/niljs/getting-started/)
 
 ## Usage
